pdf_analysis/requirements.txt
mcp>=1.2.0
pypdf>=4.2.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
rank-bm25>=0.2.2
google-generativeai>=0.7.2
numpy<2.0
